Move invariants out of the loop speculatively
If an invariant computation does not dominate all exit block of the
loop, it is generally not safe to move it outside the loop, because it
will be executed always and the instruction may raise an exception,
which would not be raised by the unoptimized program. However,
instructions, which do not raise exceptions or modify memory are safe
to execute speculatively.