Consider the entry block as having an implicit incoming executable edge
This way, values, added to the SSA worklist are not ignored. The error
scenario.

We have an entry block like:

entry:
    vZ = vX + vY
    vX = BOTTOM
    vY = BOTTOM
    ...

bX:
    vF = φ(const, vZ)
    ...

When we visit `vZ` the values of `vX` and `vY` are still `TOP`, hence
`vZ` is left as `TOP`. When we later visit `vX`, its value is lowered to
`BOTTOM` and `vX` is added to the SSA worklist. Likewise when we visit
`vY`. However, when we fetch `vZ` from the SSA worlist, it is ignored,
as the entry block has no incoming edges. Being left as `TOP`, later
`vZ` is ignored in the φ-op `vF`, which causes `vF` to be erroneously
inferred as a constant.